+++ ng/practices-ui-ktbe/src/lib/readonly-label-value/readonly-label-value.component.ts
@@ -1,4 +1,4 @@
-import { NgClass, NgIf, NgTemplateOutlet } from '@angular/common';
+import { NgClass } from '@angular/common';
 import { ChangeDetectionStrategy, Component, Input, inject } from '@angular/core';
 
 import { TranslateModule } from '@ngx-translate/core';
@@ -9,9 +9,9 @@ import { FORM_CONFIG } from '../form/form.config';
     selector: 'puibe-readonly-label-value',
     templateUrl: './readonly-label-value.component.html',
     changeDetection: ChangeDetectionStrategy.OnPush,
-    imports: [NgClass, NgIf, NgTemplateOutlet, TranslateModule],
+    imports: [NgClass, TranslateModule],
 })
-export class PuibeReadonyLabelValueComponent {
+export class PuibeReadonlyLabelValueComponent {
     readonly themeConfig = inject(FORM_CONFIG, { optional: true });
     readonly useSmallLabelFallback = this.themeConfig?.useSmallTextForReadonlyLabel;